Finish renaming files, updating includes, clean build successful
