+++ firmware/Core/Inc/telecommands/gnss_telecommand_defs.h
