+++ firmware/Core/Inc/unit_tests/unit_test_gnss.h
