+++ firmware/Core/Src/telecommands/gnss_telecommand_defs.c
