+++ firmware/Core/Src/telecommands/telecommand_definitions.c
@@ -30,7 +30,7 @@
 #include "telecommands/eps_telecommands.h"
 #include "telecommands/stm32_internal_flash_telecommand_defs.h"
 #include "telecommands/comms_telecommand_defs.h"
-#include "telecommands/gps_telecommand_defs.h"
+#include "telecommands/gnss_telecommand_defs.h"
 #include "telecommands/camera_telecommand_defs.h"
 
 #include "timekeeping/timekeeping.h"
@@ -1209,14 +1209,14 @@ const TCMD_TelecommandDefinition_t TCMD_telecommand_definitions[] = {
         .readiness_level = TCMD_READINESS_LEVEL_FLIGHT_TESTING,
     },
     // ****************** END SECTION: comms_telecommand_defs ******************
-    // ****************** SECTION: gps_telecommand_defs ******************
+    // ****************** SECTION: gnss_telecommand_defs ******************
     {
         .tcmd_name = "gnss_send_cmd_ascii",
         .tcmd_func = TCMDEXEC_gnss_send_cmd_ascii,
         .number_of_args = 1,
         .readiness_level = TCMD_READINESS_LEVEL_FOR_OPERATION,
     },
-    // ****************** END SECTION: gps_telecommand_defs ******************
+    // ****************** END SECTION: gnss_telecommand_defs ******************
     // ****************** SECTION: camera_telecommand_defs *******************
     {
         .tcmd_name = "camera_setup",
